+++ admin/views/dashboard/tmpl/default.php
@@ -4,7 +4,7 @@
  * @subpackage   Components
  * @author       Todor Iliev
  * @copyright    Copyright (C) 2016 Todor Iliev <todor@itprism.com>. All rights reserved.
- * @license      http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
 // no direct access
@@ -51,5 +51,9 @@ defined('_JEXEC') or die;
                 </tr>
             </tbody>
         </table>
+
+        <?php if (!empty($this->prismVersionLowerMessage)) {?>
+            <p class="alert alert-warning upgrade-info"><i class="icon-warning"></i> <?php echo $this->prismVersionLowerMessage; ?></p>
+        <?php } ?>
 	</div>
 </div>