+++ site/controllers/avatar.php
@@ -4,11 +4,14 @@
  * @subpackage   Components
  * @author       Todor Iliev
  * @copyright    Copyright (C) 2016 Todor Iliev <todor@itprism.com>. All rights reserved.
- * @license      http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
 // No direct access
 defined('_JEXEC') or die;
+
+jimport('Prism.libs.GuzzleHttp.init');
+jimport('Prism.libs.Aws.init');
 
 /**
  * Form controller class.