+++ site/models/contact.php
@@ -4,7 +4,7 @@
  * @subpackage   Components
  * @author       Todor Iliev
  * @copyright    Copyright (C) 2016 Todor Iliev <todor@itprism.com>. All rights reserved.
- * @license      http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
 // no direct access
@@ -100,13 +100,8 @@ class SocialCommunityModelContact extends JModelAdmin
 
             $secretKey       = $app->get('secret');
             
-            try {
-                $data['phone']   = ($data['phone'] !== null) ? Defuse\Crypto\Crypto::decrypt($data['phone'], $secretKey) : null;
-                $data['address'] = ($data['address'] !== null) ? Defuse\Crypto\Crypto::decrypt($data['address'], $secretKey) : null;
-            } catch (Exception $e) {
-                $data['phone'] = null;
-                $data['address'] = null;
-            }
+            $data['phone']   = ($data['phone'] !== null) ? Defuse\Crypto\Crypto::decrypt($data['phone'], $secretKey) : null;
+            $data['address'] = ($data['address'] !== null) ? Defuse\Crypto\Crypto::decrypt($data['address'], $secretKey) : null;
         }
 
         return $data;