+++ site/views/form/view.html.php
@@ -4,7 +4,7 @@
  * @subpackage   Components
  * @author       Todor Iliev
  * @copyright    Copyright (C) 2016 Todor Iliev <todor@itprism.com>. All rights reserved.
- * @license      http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @license      GNU General Public License version 3 or later; see LICENSE.txt
  */
 
 // no direct access
@@ -129,13 +129,8 @@ class SocialCommunityViewForm extends JViewLegacy
         $this->displayRemoveButton = 'none';
 
         $js = '
-        var socialCommunityOptions = {';
-        if (!$this->params->get('image_aspect_ratio', '')) {
-            $js .= ' aspectRatio: ""';
-        } else {
-            $js .= ' aspectRatio: ' . $this->params->get('image_aspect_ratio', '');
-        }
-        $js .= '
+        var socialCommunityOptions = {
+            aspectRatio: '. $this->params->get('image_aspect_ratio', '') . '
         }';
 
         $this->document->addScriptDeclaration($js);